+++ front/src/pages/mypage/MyBookmark.jsx
@@ -6,6 +6,7 @@ import { useEffect } from 'react';
 import axios from 'axios';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { User, bookmarkInfo } from '../../recoil/mypage';
+import MypageHeaderBtn from "../../components/buttons/mypage/MypageHeaderBtn";
 
 export default function MyBookmark() {
   const info = useRecoilValue(User);
@@ -30,22 +31,23 @@ export default function MyBookmark() {
     getData();
   }, []);
   return (
-    <div className="flex justify-center">
-      <TopSidebar />
-      <BottomSidebar />
-      <div className="flex flex-col items-center w-[50rem] h-[50rem] mt-[3rem] shadow-xss rounded-t-[2rem] bg-white">
-        {info.bookmarks.length === 0 ? <div>등록된 북마크가 없습니다.</div> : <CardList start={0} end={6} />}
-
-        {info.bookmarks.length > 6 ? (
-          <div className="shadow-pageCenter">
-            <CardList start={6} end={12} />
-            <BookmarkPagenation />
-          </div>
-        ) : null}
-
-        <div className="bg-white h-[0.5rem] w-[50rem] border-b-[1px] border-gray-300">&nbsp;</div>
-        <div className="flex bg-white h-[0.75rem] w-[50rem] rounded-[10rem]">&nbsp;</div>
+    <>
+      <MypageHeaderBtn />
+      <div className="flex justify-center">
+        <TopSidebar />
+        <BottomSidebar />
+        <div className="flex flex-col items-center w-[50rem] h-[50rem] mt-[3rem] shadow-xss rounded-t-[2rem] bg-white">
+          {info.bookmarks.length === 0 ? <div>등록된 북마크가 없습니다.</div> : <CardList start={0} end={6} />}
+          {info.bookmarks.length > 6 ? (
+            <div className="shadow-pageCenter">
+              <CardList start={6} end={12} />
+              <BookmarkPagenation />
+            </div>
+          ) : null}
+          <div className="bg-white h-[0.5rem] w-[50rem] border-b-[1px] border-gray-300">&nbsp;</div>
+          <div className="flex bg-white h-[0.75rem] w-[50rem] rounded-[10rem]">&nbsp;</div>
+        </div>
       </div>
-    </div>
+    </>
   );
 }
