+++ front/src/pages/signuppage/SignUp.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { useRecoilState, useRecoilValue } from "recoil";
 import axios from "axios";
 import { useNavigate, Navigate } from "react-router-dom";
-import { BasicCustomButton, ToSmallButton } from "../../components/Buttons";
+import { Button, ToSmallButton } from "../../components/Buttons";
 import TermsOfUse from "../../components/signuppage/TermsOfUse";
 import {
 	emailState,
@@ -197,10 +197,10 @@ export default function SignUpPage() {
 							</div>
 							<div>
 								{/* 회원가입 시도 중 버튼이 비활성화 되고, 로딩이 끝나면 다시 활성화 된다.*/}
-								<BasicCustomButton
+								<Button
 									type="submit"
 									text="Sign Up"
-									colorName="gray"
+									color="gray"
 									disabled={isLoading}
 								/>
 							</div>
