+++ front/src/pages/welcomepage/Welcome.jsx
@@ -1,16 +1,16 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable no-unused-vars */
-import { ToSmallButton } from "../../components/Buttons";
-import { useEffect } from "react";
-import axios from "axios";
-import { useRecoilState } from "recoil";
-import { authorizationTokenState } from "../../recoil/logInSignUpState";
-import { User, userInfo } from "../../recoil/mypage";
-import { currentLocationState, beadIndexState } from "../../recoil/main";
-import { locations } from "../../components/mainpage/locations";
 
-import { LogOutButton } from "../../components/buttons/LogOutButton";
-import WithdrawButton from "../../components/buttons/WithdrawButton";
+// import { useEffect } from "react";
+// import axios from "axios";
+// import { useRecoilState } from "recoil";
+// import { authorizationTokenState } from "../../recoil/logInSignUpState";
+// import { User, userInfo } from "../../recoil/mypage";
+// import { currentLocationState, beadIndexState } from "../../recoil/main";
+// import { locations } from "../../components/mainpage/locations";
+
+import { ToSmallButton } from "../../components/Buttons";
+import WithdrawButton from "../../components/buttons/mypage/WithdrawButton";
 
 export default function WelcomePage() {
 
@@ -81,16 +81,8 @@ export default function WelcomePage() {
 						<div className="text-[70px] text-[#0088F8]">MarbleUs</div>
 					</div>
 					<section className="flex gap-[40px]">
-						{/* 로그아웃, 회원탈퇴, blogButton 버튼 나중에 삭제할 것! (test중...) */}
-						<LogOutButton/>
+						{/* 회원탈퇴 버튼 나중에 삭제할 것! (test중...) */}
 						<WithdrawButton />
-						<ToSmallButton
-							linkName="blog"
-							Size="sm"
-							iconName="blog"
-							colorName="purple"
-							title="blog"
-						/>
 						<ToSmallButton
 							linkName="mainpage"
 							Size="lg"
